requirementsv1.txt

django
djangorestframework
django-cors-headers
coreapi